chore: oragnize history page logic into separate components/files
