+++ src/components/history/constants.ts
@@ -1,0 +1,6 @@
+export const DIALOG_CONTENT_CLS =
+  "bg-neutral-950 border border-white/10 shadow-2xl shadow-black/60 ring-0 max-w-sm text-neutral-100";
+export const DIALOG_CANCEL_CLS =
+  "border border-white/10 bg-neutral-900 text-neutral-300 hover:bg-white/8 hover:text-neutral-100";
+export const DIALOG_ACTION_DESTRUCTIVE_CLS =
+  "bg-red-500/15 border border-red-500/30 text-red-400 hover:bg-red-500/25 hover:text-red-300";
